AI_Internship/requirements.txt
streamlit==1.35.0
langchain==0.1.20
langchain-core==0.1.52
langchain-google-genai==0.0.8
langchain-community==0.0.38
pydantic==1.10.13
protobuf==3.20.3
faiss-cpu
PyPDF2
python-dotenv
google-generativeai==0.3.2
